requirements.txt
# =====================
# Core pipeline
# =====================
pyspark>=3.5.0
pyyaml>=6.0
pandas>=2.0

# =====================
# Database (sync + async)
# =====================
sqlalchemy>=1.4,<2.0
psycopg2-binary>=2.9
asyncpg>=0.29

# =====================
# Kafka
# =====================
kafka-python>=2.0
aiokafka>=0.10

# =====================
# Utility libraries
# =====================
python-dotenv>=1.0
loguru>=0.7

# =====================
# Dev-only tools (install with: pip install -r requirements_dev.txt)
# =====================

# pytest suite
pytest>=8.0
pytest-asyncio>=0.23

# optional formatting + linting + typing
mypy>=1.10
ruff>=0.5
black>=24.0
isort>=5.13
